Add files via upload
Develop a defensive solution for a mock organization by analyzing logs of everyday business functions to understand the environment. From these logs, we created baselines and designed custom alerts, reports, and dashboards. Additionally, we used a Splunk add-on app to monitor against a specific type of attack. On Day 1, we assessed whether the monitoring solutions we designed effectively protected VSI against attacks from JobeCorp. Finally, we presented a summary of the work you completed.